Convert to a PlatformIO project

+++ include/AirTime.h
@@ -6,7 +6,7 @@
 
 constexpr uint8_t SECONDS_PER_MINUTE = 60;
 constexpr uint8_t MINUTES_PER_HOUR = 60;
-constexpr uint16_t MS_PER_MINUTE = SECONDS_PER_MINUTE * 1000;
+constexpr uint16_t MS_PER_MINUTE = (uint16_t)SECONDS_PER_MINUTE * 1000;
 constexpr uint32_t MS_PER_HOUR =
     static_cast<uint32_t>(MINUTES_PER_HOUR) * MS_PER_MINUTE;
